modules/bilibili_client.py
import asyncio
import websockets
import json
import threading
from queue import Queue
import time


class BilibiliDanmakuClient:
    def __init__(self, room_id, message_queue):
        self.room_id = room_id
        self.message_queue = message_queue
        self.running = False

    async def connect(self):
        """连接到B站弹幕服务器"""
        uri = f"wss://broadcastlv.chat.bilibili.com/sub"

        # 这里需要实际的连接逻辑
        # 简化版：使用第三方库 bilibili-api 的实现
        await self._connect_bilibili()

    async def _connect_bilibili(self):
        """实际的连接逻辑"""
        # 注意：这里需要安装 bilibili-api 库
        # pip install bilibili-api-python
        from bilibili_api import sync, live

        try:
            room = live.LiveRoom(room_display_id=int(self.room_id))

            @room.on('DANMU_MSG')
            async def on_danmaku(event):
                """处理弹幕消息"""
                info = event['data']['info']
                user = info[2][1]  # 用户名
                text = info[1]  # 弹幕内容

                # 将弹幕放入消息队列
                message_data = {
                    'type': 'danmaku',
                    'user': user,
                    'text': text,
                    'timestamp': time.time()
                }
                self.message_queue.put(message_data)
                print(f"收到弹幕: {user}: {text}")

            # 开始监听
            await room.connect()

        except Exception as e:
            print(f"连接B站失败: {e}")

    def start(self):
        """启动弹幕监听"""
        self.running = True
        thread = threading.Thread(target=self._run_async)
        thread.daemon = True
        thread.start()

    def _run_async(self):
        """在新线程中运行异步代码"""
        loop = asyncio.new_event_loop()
        asyncio.set_event_loop(loop)
        loop.run_until_complete(self.connect())