requirements.txt
websocket-client
requests
google-generativeai
pygame
asyncio
threading
queue
json
time
datetime
bilibili-api-python